Add files via upload
IMDB is a popular movie-related search engine. It provides a sizable database of movies, actors, and directors, as well as details on each movie and reviews. Because of its popularity, people use it to find out more about their favourite actors and films. The IMDB website, however, is challenging to use and doesn't always provide the most accurate information about a particular actor or film. If one wants to build a recommendation engine for movies that suggests movies based on user preferences, one needs data sets of various movies and TV shows from various genres. It is possible for programs to automatically extract all of this data from IMDB and then use models to analyze it. This makes it possible to get the results you want. Keywords: Sentiment analysis, Keyword Extraction, LDA